fix: 유저 프롬프트 입력 오류 해결 (#29) - 오타 수정 - product product_id INteger -> BIGINT로 변경 상품명이 너무 길어서 저장이 안됨 추후에 VARCHAR로 변경하는 부분도 생각해보면 좋을 거 같습니다. 급하게 한다고 일단 BIGINT로 했습니다. - crud로직에 DB에 저장을 하고 Return이 안되서 If문을 수행할때 실행하지 못하는 오류가 있어서 해결했습니다 -> 오래 걸렸습니다 이거 - 나머지 로직상 dictionaty를 Class이름 빼는거 처럼 사용하고 있어서 그 부분도 해결했습니다.

+++ models/image.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, func, ForeignKey
+from sqlalchemy import Column, Integer, String, DateTime, func, ForeignKey, BIGINT
 from sqlalchemy.orm import relationship
 from db.database import base
 
@@ -7,7 +7,7 @@ class Image(base):
 
     image_id = Column(Integer, primary_key=True, index=True)
     img_url = Column(String(255), nullable=False)
-    product_id = Column(Integer, ForeignKey("product.product_id"),nullable=False)
+    product_id = Column(BIGINT, ForeignKey("product.product_id"),nullable=False)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     updated_at = Column(DateTime(timezone=True), server_default=func.now(), onupdate=func.now())
